Warnings closed.

+++ u3d_robot_module.cpp
@@ -211,7 +211,7 @@ void *u3dRobotModule::writePC(unsigned int *buffer_length) {
 }
 
 FunctionResult* u3dRobot::executeFunction(system_value functionId, void **args) {
-	if ((functionId < 1) || (functionId > COUNT_u3dRobot_FUNCTIONS)) {
+	if ((functionId < 1) || (functionId > (int) COUNT_u3dRobot_FUNCTIONS)) {
 		return NULL;
 	}
 	variable_value rez = 0;
